+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # v2.7.40
 
+* more bulletproof `SelfCleaningDefaultDict`